backend/requirements.txt
# requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.10.0
httpx==0.27.0
pydantic==2.9.2
requests==2.32.0
python-multipart==0.0.9
websocket-client==1.8.0
openai==2.6.1